toolchain/*: Use EXCLUDE_FROM_ALL 1
This can help ignore fuck ups that try require re-building the toolchain.
Rare but annoying and this is better than editing a few files to run `true` rather than `make` or pointlessly rebuilding the whole toolchain again.

Just test this for now, obviously any toolchain updates or removing the prefix will require explicit `ninja gcc` to build the toolchain but this is the case anyway due to how the deps are on packages/*