README: Update build deps (ninja -> samu)
Replace ninja with samurai, I've been using it fine for _everything_ for the past 1.5 year so I'll add it here as a default. Technically this project isn't tested with ninja.

It's a direct replacement (alpine symlinks ninja -> samu), so no changes in build commands needed and it will work perfectly fine with regular ninja installed instead.

It is cheeky (hence why I put it off for so long) since it could replace ninja on your machine but this project is supposed to be ran as a container (wsl2 or an nspawn container) so not my problem really.