An exercise in debugging freetard software build errors...
This is amazing.
Since moving back to GCC I haven't been able to build ffmpeg with dav1d decoder
because pkg-cnfig fails to find it, the ffmpeg config.log shows that it does
but is skipped for being "incompatible" I spent hours trying to figure out what
was wrong, almost as if somehow with no changes it was not really 64bit. So I
had to use the inferior aom decoder, the problem with that was that unless I
explicitly told it to use it's own toolchain file ffmpeg would fail to
compile/link. After looking a hundred times I noticed that the aom toolchain
file sets CMAKE_SYSTEM_PROCESSOR to x86_64 but for some reason cmake on the
machine sets it to x86 when generating the toolchain file. Fixing this little
error (hopefully) has allowed ffmpeg to compile when aom is using
${TOOLCHAIN_FILE} it has also magically allowed ffmpeg to find dav1d and now
fully builds and link as expected. Dav1d uses meson, ffmpeg uses autotools,
only aom uses cmake.

I have no idea how, or why this happened but it's just amazing. I've been
wracking my brains out over this for almost a week even looking in to the toolchain itself trying to find how it could be wrong.